[Refactor] 월간 달력 헤더, 내용 분리 기존에는 모든 파트가 붙어있었는데 분리하였습니다.
